nope, actually fix it this time

+++ Assets/__Scripts/Extensions/ColorExtensions.cs
@@ -2,7 +2,7 @@
 
 static class ColorExtensions
 {
-    public static Color WithAlpha(this Color color, int alpha)
+    public static Color WithAlpha(this Color color, float alpha)
     {
         return new Color(color.r, color.g, color.b, alpha);
     }
